requirements.txt
mecab-python3
transformers
torch
tqdm
fugashi
ipadic
datasets
unidic-lite
accelerate 